Revert "Changes from https://github.com/EvanBacon/expo-apple-targets/pull/16 …"
This reverts commit a224e3ce762270d41dedac79c45d50bf420913d6.
