+++ apps/app-clip-demo/App.js
@@ -1,3 +1,4 @@
+import React from "react";
 import { StyleSheet, Text, View } from "react-native";
 
 export default function App() {
