+++ apps/kitchen/package.json
@@ -10,10 +10,10 @@
   },
   "dependencies": {
     "@bacons/apple-targets": "*",
-    "expo": "^54",
-    "expo-splash-screen": "~31.0.11",
-    "react": "19.1.0",
-    "react-native": "0.81.5"
+    "expo": "^53",
+    "expo-splash-screen": "~0.30.8",
+    "react": "19.0.0",
+    "react-native": "0.79.2"
   },
   "private": true
 }
