+++ packages/apple-targets/package.json
@@ -1,10 +1,9 @@
 {
   "name": "@bacons/apple-targets",
-  "version": "3.0.6",
+  "version": "3.0.5",
   "description": "Generate Apple Targets with Expo Prebuild",
   "main": "build/ExtensionStorage.js",
   "types": "build/ExtensionStorage.d.ts",
-  "homepage": "https://github.com/evanbacon/expo-apple-targets",
   "files": [
     "app.plugin.js",
     "app.plugin.d.ts",
@@ -31,11 +30,7 @@
     "config-plugins",
     "apple-targets"
   ],
-  "author": {
-    "email": "baconbrix@gmail.com",
-    "name": "Evan Bacon",
-    "url": "https://evanbacon.dev"
-  },
+  "author": "Evan Bacon",
   "license": "MIT",
   "dependencies": {
     "@bacons/xcode": "1.0.0-alpha.27",
