+++ packages/apple-targets/src/config.ts
@@ -104,6 +104,8 @@ export type Config = {
 
   /**
    * A local file path or URL to an image asset.
+   * @example "./assets/icon.png"
+   * @example "https://example.com/icon.png"
    */
   icon?: string;
