+++ packages/apple-targets/src/icon/withImageAsset.ts
