+++ packages/apple-targets/src/withEasCredentials.ts
@@ -3,7 +3,7 @@ import { ConfigPlugin } from "expo/config-plugins";
 
 import { Entitlements } from "./config";
 import { getAuxiliaryTargets, getMainAppTarget } from "./target";
-import { withXcodeProjectBeta } from "./with-bacons-xcode";
+import { withXcodeProjectBeta } from "./withXcparse";
 
 const debug = require("debug")("expo:target:eas") as typeof console.log;
