+++ packages/apple-targets/src/withWidget.ts
@@ -5,26 +5,18 @@ import { globSync } from "glob";
 import path from "path";
 import chalk from "chalk";
 
-import { withIosColorset } from "./colorset/with-ios-colorset";
+import { withIosColorset } from "./colorset/withIosColorset";
 import { Config, Entitlements } from "./config";
-import { withImageAsset } from "./icon/with-image-asset";
-import { withIosIcon } from "./icon/with-ios-icon";
+import { withImageAsset } from "./icon/withImageAsset";
+import { withIosIcon } from "./icon/withIosIcon";
 import {
   getFrameworksForType,
   getTargetInfoPlistForType,
   getWatchAppTarget,
   SHOULD_USE_APP_GROUPS_BY_DEFAULT,
 } from "./target";
-import { withEASTargets } from "./with-eas-credentials";
-import { withXcodeChanges } from "./with-xcode-changes";
-import {
-  getSanitizedBundleIdentifier,
-  LOG_QUEUE,
-  logOnce,
-  sanitizeNameForNonDisplayUse,
-  warnOnce,
-} from "./util";
-import type { DeviceFamily } from "./configuration-list";
+import { withEASTargets } from "./withEasCredentials";
+import { DeviceFamily, withXcodeChanges } from "./withXcodeChanges";
 
 type Props = Config & {
   directory: string;
@@ -33,8 +25,43 @@ type Props = Config & {
 
 const DEFAULT_DEPLOYMENT_TARGET = "18.0";
 
+function memoize<T extends (...args: any[]) => any>(fn: T): T {
+  const cache = new Map<string, any>();
+  return ((...args: any[]) => {
+    const key = JSON.stringify(args);
+    if (cache.has(key)) {
+      return cache.get(key);
+    }
+    const result = fn(...args);
+    cache.set(key, result);
+    return result;
+  }) as T;
+}
+
+const warnOnce = memoize(console.warn);
+const logOnce = memoize(console.log);
+
+function createLogQueue(): { add: (fn: Function) => void; flush: () => void } {
+  const queue: Function[] = [];
+
+  const flush = () => {
+    queue.forEach((fn) => fn());
+    queue.length = 0;
+  };
+
+  return {
+    flush,
+    add: (fn: Function) => {
+      queue.push(fn);
+    },
+  };
+}
+
+// Queue up logs so they only run when prebuild is actually running and not during standard config reads.
+const prebuildLogQueue = createLogQueue();
+
 const withWidget: ConfigPlugin<Props> = (config, props) => {
-  LOG_QUEUE.add(() =>
+  prebuildLogQueue.add(() =>
     warnOnce(
       chalk`\nUsing experimental Config Plugin {bold @bacons/apple-targets} that is subject to breaking changes.`
     )
@@ -184,7 +211,7 @@ const withWidget: ConfigPlugin<Props> = (config, props) => {
         if (Array.isArray(mainAppGroups) && mainAppGroups.length > 0) {
           // Then set the target app groups to match the main app.
           entitlements[APP_GROUP_KEY] = mainAppGroups;
-          LOG_QUEUE.add(() => {
+          prebuildLogQueue.add(() => {
             logOnce(
               chalk`[${targetDirName}] Syncing app groups with main app. {dim Define entitlements[${JSON.stringify(
                 APP_GROUP_KEY
@@ -192,7 +219,7 @@ const withWidget: ConfigPlugin<Props> = (config, props) => {
             );
           });
         } else {
-          LOG_QUEUE.add(() =>
+          prebuildLogQueue.add(() =>
             warnOnce(
               chalk`{yellow [${targetDirName}]} Apple target may require the App Groups entitlement but none were found in the Expo config.\nExample:\n${JSON.stringify(
                 {
@@ -259,12 +286,12 @@ const withWidget: ConfigPlugin<Props> = (config, props) => {
   withDangerousMod(config, [
     "ios",
     async (config) => {
-      LOG_QUEUE.flush();
+      prebuildLogQueue.flush();
 
       fs.mkdirSync(targetDirAbsolutePath, { recursive: true });
 
       const files: [string, string][] = [
-        ["Info.plist", plist.build(getTargetInfoPlistForType(props.type))],
+        ["Info.plist", getTargetInfoPlistForType(props.type)],
       ];
 
       // if (props.type === "widget") {
@@ -417,3 +444,18 @@ const withConfigColors: ConfigPlugin<Pick<Props, "colors" | "directory">> = (
 };
 
 export default withWidget;
+
+function getSanitizedBundleIdentifier(value: string) {
+  // According to the behavior observed when using the UI in Xcode.
+  // Must start with a letter, period, or hyphen (not number).
+  // Can only contain alphanumeric characters, periods, and hyphens.
+  // Can have empty segments (e.g. com.example..app).
+  return value.replace(/(^[^a-zA-Z.-]|[^a-zA-Z0-9-.])/g, "-");
+}
+
+function sanitizeNameForNonDisplayUse(name: string) {
+  return name
+    .replace(/[\W_]+/g, "")
+    .normalize("NFD")
+    .replace(/[\u0300-\u036f]/g, "");
+}
