+++ packages/create-target/package.json
@@ -1,6 +1,6 @@
 {
   "name": "create-target",
-  "version": "2.0.1",
+  "version": "2.0.0",
   "description": "Create targets for your Expo app",
   "bin": "./build/index.js",
   "repository": {
@@ -8,7 +8,6 @@
     "url": "https://github.com/evanbacon/expo-apple-targets.git",
     "directory": "packages/create-target"
   },
-  "homepage": "https://github.com/evanbacon/expo-apple-targets",
   "files": [
     "build"
