+++ packages/create-target/src/createAsync.ts
@@ -2,7 +2,6 @@
 import chalk from "chalk";
 import fs from "fs";
 import path from "path";
-import plist from "@expo/plist";
 
 import { normalizeStaticPlugin } from "@expo/config-plugins/build/utils/plugin-resolver";
 import { ExpoConfig, getConfig, modifyConfigAsync } from "@expo/config";
@@ -154,7 +153,7 @@ export async function createAsync(
   Log.log(chalk`Writing {cyan Info.plist} file`);
   await fs.promises.writeFile(
     path.join(targetDir, "Info.plist"),
-    plist.build(getTargetInfoPlistForType(resolvedTarget as any))
+    getTargetInfoPlistForType(resolvedTarget as any)
   );
 
   Log.log(
