--- packages/create-target/templates/network-filter-data/FilterDataProvider.swift
@@ -1,19 +1,0 @@
-import NetworkExtension
-
-class FilterDataProvider: NEFilterDataProvider {
-
-    override func startFilter(completionHandler: @escaping (Error?) -> Void) {
-        // Add code to initialize the filter.
-        completionHandler(nil)
-    }
-    
-    override func stopFilter(with reason: NEProviderStopReason, completionHandler: @escaping () -> Void) {
-        // Add code to clean up filter resources.
-        completionHandler()
-    }
-    
-    override func handleNewFlow(_ flow: NEFilterFlow) -> NEFilterNewFlowVerdict {
-        // Add code to determine if the flow should be dropped or not, downloading new rules if required.
-        return .allow()
-    }
-}
